ppu scrolling actually actually sort of working

+++ SiroGen83/vector2.h
@@ -1,4 +1,4 @@
 struct Vector2 {
-	float x;
-	float y;
+	unsigned short x;
+	unsigned short y;
 };
